% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicdata}
\alias{basicdata}
\title{Example Dataset for a Survival Outcome with Competing Events}
\format{
A data table with 11,332 rows and 8 variables:
\describe{
  \item{t0}{Time index.}
  \item{id}{Unique identifier for each individual.}
  \item{L1}{Binary time-varying covariate.}
  \item{L2}{Continuous time-varying covariate.}
  \item{L3}{Continuous baseline covariate. For each individual, the baseline values are repeated at each time point.}
  \item{A}{Binary treatment variable.}
  \item{D}{Competing event; time-varying indicator of failure.}
  \item{Y}{Outcome of interest; time-varying indicator of failure.}
}
}
\usage{
basicdata
}
\description{
A dataset consisting of 11,332 observations on 2,500 individuals over 7 time points. Each row in the dataset corresponds to the record of one individual at one time point. Individuals who are censored at time \eqn{k+1} only have a total of \eqn{k+1} records, which correspond to time indices \eqn{0,..., k}.
}
\keyword{datasets}
