% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicdata_nocomp}
\alias{basicdata_nocomp}
\title{Example Dataset for a Survival Outcome without Censoring}
\format{
A data table with 13,170 rows and 7 variables:
\describe{
  \item{t0}{Time index.}
  \item{id}{Unique identifier for each individual.}
  \item{L1}{Binary covariate.}
  \item{L2}{Continuous covariate.}
  \item{L3}{Continuous baseline covariate. For each individual, the baseline values are repeated at each time point.}
  \item{A}{Binary treatment variable.}
  \item{Y}{Outcome of interest; time-varying indicator of failure.}
}
}
\usage{
basicdata_nocomp
}
\description{
A dataset consisting of 13,170 observations on 2,500 individuals over 7 time points. Each row in the dataset corresponds to the record of one individual at one time point.
}
\keyword{datasets}
