% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censor_data}
\alias{censor_data}
\title{Example Dataset for a Survival Outcome with an Indicator of Censoring Variable}
\format{
A data table with 22,500 rows and 7 variables:
\describe{
  \item{t0}{Time index.}
  \item{id}{Unique identifier for each individual.}
  \item{L}{Binary time-varying covariate.}
  \item{A}{Continuous treatment variable.}
  \item{C}{Censoring indicator.}
  \item{Y}{Outcome of interest; time-varying indicator of failure.}
}
}
\usage{
censor_data
}
\description{
A dataset consisting of 118,725 observations on 20,000 individuals over 10 time points. Each row in the dataset corresponds to the record of one individual at one time point.
}
\keyword{datasets}
