% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continuous_eofdata_pb}
\alias{continuous_eofdata_pb}
\title{Example Dataset for a Continuous Outcome at End of Follow-Up with Pre-Baseline Times}
\format{A data table with 22,500 rows and 7 variables:
\describe{
  \item{t0}{Time index.}
  \item{id}{Unique identifier for each individual.}
  \item{L1}{Categorical time-varying covariate.}
  \item{L2}{Continuous time-varying covariate.}
  \item{L3}{Continuous baseline covariate. For each individual, the baseline values are repeated at each time point.}
  \item{A}{Binary treatment variable.}
  \item{Y}{Continuous outcome of interest. Because this outcome is only defined at the end of follow-up, values of \code{NA} are given in all other time points.}
}}
\usage{
continuous_eofdata_pb
}
\description{
A dataset consisting of 22,500 observations on 2,500 individuals over 2 pre-baseline time points and follow-up 7 time points. Each row in the dataset corresponds to the record of one individual at one time point.
}
\keyword{datasets}
