% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{example_intervention_1}
\alias{example_intervention_1}
\title{Example Intervention 1}
\usage{
example_intervention_1(newdf, pool, intvar, intvals, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{intvar}{Character string specifying the name of the variable to be intervened on in each round of the simulation.}

\item{intvals}{A list. The first entry is user-specified cutoff for the time variable, above
which \code{intvar} will be assigned some user-specified
value and below which the intervention variable will be assigned some
other user-specified value. The second and third entries are the values
assigned to \code{intvar} if the time value is above and below, respectively, the cutoff.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function implements an example intervention for the specified intervention variable in
the data table \code{newdf}. It allows the user to define a cutoff for the time variable, where a
different intervention is assigned depending on whether the time is before or after the cutoff.
The data table \code{newdf} is modified in place.
}
