% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{example_intervention_2}
\alias{example_intervention_2}
\title{Example Intervention 2}
\usage{
example_intervention_2(newdf, pool, intvar, intvals, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{intvar}{Character string specifying the name of the variable to be intervened on in each round of the simulation.}

\item{intvals}{A list of length 1. The entry in the list is a user-specified value for the variable \code{A}, such that
if the simulated value of \code{A} is equal to some value, \code{intvar} is set equal
to 1, while otherwise the intervention variable is set equal to 0.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function implements an example intervention for the specified intervention variable in
the data table \code{newdf}. It allows the user to override the natural course, such that if the
simulated natural course treatment is equal to some user-defined value, the treatment
variable is set to 1, and otherwise the treatment variable is set to 0. The data table
\code{newdf} is modified in place.
}
