% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit GLM on Covariate}
\usage{
fit_glm(covparams, covlink = NA, covfam, obs_data, j)
}
\arguments{
\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.).}

\item{covlink}{Vector of link functions.}

\item{covfam}{Vector of character strings specifying the names of the family functions to be
used for fitting the GLM.}

\item{obs_data}{Data on which the model is fit.}

\item{j}{Integer specifying the index of the covariate.}
}
\value{
Fitted model for the covariate at index \eqn{j}.
}
\description{
This internal function fits a generalized linear model (GLM) for a single covariate using the observed data.
}
\keyword{internal}
