% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{fit_multinomial}
\alias{fit_multinomial}
\title{Fit Multinomial Model on Covariate}
\usage{
fit_multinomial(covparams, obs_data, k)
}
\arguments{
\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.).}

\item{obs_data}{Data on which the model is fit.}

\item{k}{Integer specifying the index of the covariate.}
}
\value{
Fitted model for the covariate at index \eqn{k}.
}
\description{
This internal function fits a multinomial regression model for a categorical covariate using the observed data.
}
