% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{fit_trunc_normal}
\alias{fit_trunc_normal}
\title{Fit Truncated Normal Model on Covariate}
\usage{
fit_trunc_normal(covparams, obs_data, j)
}
\arguments{
\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.).}

\item{obs_data}{Data on which the model is fit.}

\item{j}{Integer specifying the index of the covariate.}
}
\value{
Fitted model for the covariate at index \eqn{j}.
}
\description{
This internal function models a covariate using a normal distribution truncated on one
side at a user-specified cutoff.
}
