% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{get_outgrphs}
\alias{get_outgrphs}
\title{Get Risk and Survival Plots}
\usage{
get_outgrphs(x, risk, survival, xlab, ylab_risk, ylab_surv, ci_risk)
}
\arguments{
\item{x}{Object of class "gformula_survival".}

\item{risk}{Logical scalar indicating whether to include a plot for the risk.}

\item{survival}{Logical scalar indicating whether to include a plot for the survival.}

\item{xlab}{Character string for the x axes of all plots.}

\item{ylab_risk}{Character string for the y axis of the plot for the risk (if applicable).}

\item{ylab_surv}{Character string for the y axis of the plot for the survival (if applicable).}

\item{ci_risk}{Logical scalar specifying whether to include error bars for the 95\% confidence intervals of the estimated risk under the natural course. This argument is only effective if the argument \code{nsamples} was set to a positive value in \code{\link{gformula_survival}}.}
}
\value{
A list of plots for the risk and survival.
}
\description{
This internal function obtains the risk and survival plots for \code{\link{gformula_survival}}.
}
\keyword{internal}
