% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{gformula_continuous_eof}
\alias{gformula_continuous_eof}
\title{Estimation of Continuous End-of-Follow-Up Outcome Under the Parametric G-Formula}
\usage{
gformula_continuous_eof(obs_data, id = "id", time_points,
  time_name = "t0", covnames, covtypes, covparams, covfits_custom = NA,
  covpredict_custom = NA, histvars = NA, histories = NA,
  basecovs = NA, outcome_name = "Y", ymodel, intvars = NA,
  interventions = NA, int_descript = NA, ref_int = 0,
  visitprocess = NA, restrictions = NA, yrestrictions = NA,
  nsimul = NA, sim_data_b = FALSE, seed, nsamples = 0,
  parallel = FALSE, ncores = NA, ...)
}
\arguments{
\item{obs_data}{Data table containing the observed data.}

\item{id}{Character string specifying the name of the ID variable in \code{obs_data}. The default is \code{"id"}.}

\item{time_points}{Number of time points to simulate.}

\item{time_name}{Character string specifying the name of the time variable in \code{obs_data}. The default is \code{"t0"}.}

\item{covnames}{Vector of character strings specifying the names of the time-varying covariates in \code{obs_data}.}

\item{covtypes}{Vector of character strings specifying the "type" of each time-varying covariate included in \code{covnames}. The possible "types" are: \code{"binary"}, \code{"normal"}, \code{"categorical"}, \code{"bounded normal"}, \code{"zero-inflated normal"}, \code{"truncated normal"}, and \code{"absorbing"}.}

\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.). Each vector
must be the same length as \code{covnames} and in the same order.
If a parameter is not required for a certain covariate, it
should be set to \code{NA} at that index.}

\item{covfits_custom}{Vector containing custom fit functions for time-varying covariates that
do not fall within the pre-defined covariate types. It should be in
the same order \code{covnames}. If a custom fit function is not
required for a particular covariate (e.g., if the first
covariate is of type \code{"binary"} but the second is of type \code{"custom"}), then that
index should be set to \code{NA}. The default is \code{NA}.}

\item{covpredict_custom}{Vector containing custom prediction functions for time-varying
covariates that do not fall within the pre-defined covariate types.
It should be in the same order as \code{covnames}. If a custom
prediction function is not required for a particular
covariate, then that index should be set to \code{NA}. The default is \code{NA}.}

\item{histvars}{Vector of character strings specifying the names of the variables for which history functions
are to be applied. The default is \code{NA}.}

\item{histories}{Vector of history functions to apply to the variables specified in \code{histvars}. The default is \code{NA}.}

\item{basecovs}{Vector of character strings specifying the names of baseline covariates in \code{obs_data}. These covariates are not simulated using a model but rather carry their value over all time points from the first time point of \code{obs_data}. These covariates should not be included in \code{covnames}. The default is \code{NA}.}

\item{outcome_name}{Character string specifying the name of the outcome variable in \code{obs_data}. The default is \code{"Y"}.}

\item{ymodel}{Model statement for the outcome variable.}

\item{intvars}{Vector of character strings specifying the names of the variables to be intervened
on in each round of the simulation. The default is \code{NA}.}

\item{interventions}{List of vectors. Each vector contains a function
implementing a particular intervention, optionally
followed by one or more "intervention values" (i.e.,
integers used to specify the treatment regime).
The default is \code{NA}.}

\item{int_descript}{Vector of character strings, each describing an intervention. It must
be in same order as the entries in \code{interventions} and should not include a description for the natural course.}

\item{ref_int}{Integer denoting the intervention to be used as the
reference for calculating the end-of-follow-up mean ratio. 0 denotes the
natural course, while subsequent integers denote user-specified
interventions in the order that they are
named in \code{interventions}. The default is 0.}

\item{visitprocess}{List of vectors. Each vector contains as its first entry
the covariate name of a visit process; its second entry
the name of a covariate whose modeling depends on the
visit process; and its third entry the maximum number
of consecutive visits that can be missed before an
individual is censored. The default is \code{NA}.}

\item{restrictions}{List of vectors. Each vector contains as its first entry
the covariate affected by the restriction; its second entry
the condition that must be \code{TRUE} for the covariate to be
modeled; its third entry a function that executes other
specific actions based on the condition; and its fourth
entry some value used by the function. The default is \code{NA}.}

\item{yrestrictions}{List of vectors. Each vector containins as its first entry
a condition and its second entry an integer. When the
condition is \code{TRUE}, the outcome variable is simulated
according to the fitted model; when the condition is \code{FALSE},
the outcome variable takes on the value in the second entry.
The default is \code{NA}.}

\item{nsimul}{Number of subjects for whom to simulate data. By default, this argument is set
equal to the number of subjects in \code{obs_data}.}

\item{sim_data_b}{Logical scalar indicating whether to return the simulated data set. If bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0), this argument must be set to \code{FALSE}. The default is \code{FALSE}.}

\item{seed}{Starting seed for simulations and bootstrapping.}

\item{nsamples}{Integer specifying the number of bootstrap samples to generate.
The default is 0.}

\item{parallel}{Logical scalar indicating whether to parallelize simulations of
different interventions to multiple cores.}

\item{ncores}{Integer specifying the number of cores to use in parallel
simulation.}

\item{...}{Other arguments, which are passed to the functions in \code{covpredict_custom}.}
}
\value{
An object of class \code{gformula_continuous_eof}. The object is a list with the following components:
\item{result}{Results table containing the estimated mean outcome for all interventions (inculding natural course) at the last time point. If bootstrapping was used, the results table includes the bootstrap end-of-follow-up mean ratio, standard error, and 95\% confidence interval.}
\item{coeffs}{A list of the coefficients of the fitted models.}
\item{rmses}{A list of root mean square error (RMSE) values of the fitted models.}
\item{sim_data}{A list of data tables of the simulated data. Each element in the list corresponds to one of the interventions. If the argument \code{sim_data_b} is set to \code{FALSE}, a value of \code{NA} is given.}
\item{...}{Some additional elements.}

The results for the g-formula simulation under various interventions for the last time point are printed with the \code{\link{print.gformula_continuous_eof}} function. To generate graphs comparing the mean estimated and observed covariate values over time, use the \code{\link{print.gformula_continuous_eof}} function.
}
\description{
Based on an observed data set, this function estimates the outcome mean at end-of-follow-up under
multiple user-specified interventions using the parametric g-formula. See the documentation of (TO DO - link to vignette) for
further details concerning the application and implementation of the parametric g-formula.
}
\examples{
## TO ADD AFTER PAPER IS FINALIZED

}
\references{
Robins JM. A new approach to causal inference in mortality studies with a sustained exposure period: application to the healthy worker survivor effect. Mathematical Modelling. 1986;7:1393–1512. [Errata (1987) in Computers and Mathematics with Applications 14, 917.-921. Addendum (1987) in Computers and Mathematics with Applications 14, 923-.945. Errata (1987) to addendum in Computers and Mathematics with Applications 18, 477.].
}
