% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hr_helper}
\alias{hr_helper}
\title{Format Simulated Dataset for Hazard Ratio Calculation}
\usage{
hr_helper(i, intcomp, time_name, pools)
}
\arguments{
\item{i}{Integer specifying the index \code{intcomp}.}

\item{intcomp}{List of two numbers indicating a pair of interventions to be compared by a hazard ratio.}

\item{time_name}{Character string specifying the name of the time variable in \code{pools}.}

\item{pools}{Data table containing the simulated values for the covariates, outcome
probabilities, competing event probabilities, outcomes, and competing
events.}
}
\value{
Data table consisting of failure time information for each individual or the last time
                   point information (for individuals who do not experience an event).
}
\description{
This internal function modifies the pooled-over-time dataset generated by the \code{\link{simulate}} function
to a format suitable for hazard ratio calculation.
}
