% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histories.R
\name{lagged}
\alias{lagged}
\alias{cumavg}
\alias{lagavg}
\title{History functions}
\usage{
lagged(pool, histvars, time_name, t, id, max_visits)

cumavg(pool, histvars, time_name, t, id, max_visits)

lagavg(pool, histvars, time_name, t, id, max_visits)
}
\arguments{
\item{pool}{Data table containing all information prior to time \eqn{t} (\eqn{t} noninclusive).}

\item{histvars}{Vector of character strings specifying the names of the variables for which history functions are to be applied.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool}.}

\item{t}{Integer specifying the current time index.}

\item{id}{Character string specifying the name of the ID variable in \code{pool}.}

\item{max_visits}{This argument is not applicable in these functions.}
}
\description{
These functions create new columns in an input data table for covariate histories. Users must specify which covariates are to be used in the history functions. The data table is modified in place.
}
\details{
\code{lagged} creates new columns for lagged versions of existing
variables in the dataset. The user must specify which variables are to be lagged.

\code{cumavg} creates new columns for the cumulative average up until
time \eqn{t} of existing variables in the dataset.

\code{lagavg} creates new columns for the "lagged cumulative average"
(cumulative average up until time t, then lagged by one time unit) up until time \eqn{t} of existing
variables in the dataset.
}
