% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histories.R
\name{make_histories}
\alias{make_histories}
\title{Generates Functions of History of Existing Covariates}
\usage{
make_histories(pool, histvars, histories, time_name, t, id, max_visits)
}
\arguments{
\item{pool}{Data table containing all information prior to time \eqn{t} (\eqn{t} noninclusive).}

\item{histvars}{Vector of character strings specifying the names of the variables for which history functions are to be applied.}

\item{histories}{Vector of history functions to apply to the variables specified in \code{histvars}.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool}.}

\item{t}{Integer specifying the current time index.}

\item{id}{Character string specifying the name of the ID variable in \code{pool}.}

\item{max_visits}{A vector of one or more values denoting the maximum number of times
a binary covariate representing a visit process may be missed before
the individual is censored from the data (in the observed data) or
a visit is forced (in the simulated data). Multiple values exist in the
vector when the modeling of more than covariate is attached to a visit
process. A value of \code{NA} should be provided when there is no visit process.}
}
\value{
A data table containing all preexisting information at time \eqn{t}, plus columns containing
                  new function-of-history variables.
}
\description{
This internal function applies the history functions to create new columns in an input data table containing new variables that are functions
of the histories of existing variables in the dataset.
}
