% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{natural}
\alias{natural}
\title{Natural Course Intervention}
\usage{
natural(newdf, pool, intvar, intvals, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{intvar}{Character string specifying the name of the variable to be intervened on in each round of the simulation.}

\item{intvals}{A list of user-specified values for the intervention.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function implements the natural course (i.e., model-based simulation) for the specified
intervention variable in the data table newdf. Because \code{newdf} is initiated with the
natural course, this function does nothing.
}
