% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.gformula_binary_eof}
\alias{plot.gformula_binary_eof}
\title{Plot method for objects of class "gformula_binary_eof"}
\usage{
\method{plot}{gformula_binary_eof}(x, covnames = NULL, ncol = NULL,
  nrow = NULL, common.legend = TRUE, legend = "bottom",
  xlab = NULL, ylab_cov = NULL, ...)
}
\arguments{
\item{x}{Object of class "gformula_binary_eof".}

\item{covnames}{Vector of character strings specifying the names of the time-varying covariates to be plotted. The ordering of covariates given here is used in the plot grid. Time-varying covariates of type \code{"categorical time"} cannot be included. By default, this argument is set equal to the \code{covnames} argument used in \code{\link{gformula_binary_eof}}, where covariates of type \code{"categorical time"} are removed.}

\item{ncol}{Number of columns in the plot grid. By default, two columns are used when there is at least two plots.}

\item{nrow}{Number of rows in the plot grid. By default, a maximum of six rows is used and additional plots are included in subsequent pages.}

\item{common.legend}{Logical scalar indicating whether to include a legend. The default is \code{TRUE}.}

\item{legend}{Character string specifying the legend position. Valid values are \code{"top"}, \code{"bottom"}, \code{"left"}, \code{"right"}, and \code{"none"}. The default is \code{"bottom"}.}

\item{xlab}{Character string for the x axes of all plots. By default, this argument is set to the \code{time_name} argument specified in \code{\link{gformula_binary_eof}}.}

\item{ylab_cov}{Vector of character strings for the y axes of the plots for the covariates. This argument must be the same length as \code{covnames}. The i-th element of this argument corresponds to the plot for the i-th element of \code{covnames}.}

\item{...}{Other arguments, which are passed to \code{\link[ggpubr]{ggarrange}}.}
}
\description{
This function generates graphs of the mean simulated vs. observed values at each time point of the
time-varying covariates under the natural course. For categorical covariates,
the observed and simulated counts of the levels of the factors are plotted at each time point.
}
\seealso{
\code{\link{gformula_binary_eof}}
}
