% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_fun_D}
\alias{pred_fun_D}
\title{Fit Competing Event Model}
\usage{
pred_fun_D(model, compevent_restrictions, obs_data)
}
\arguments{
\item{model}{Model statement for competing event variable.}

\item{compevent_restrictions}{List of vectors. Each vector containins as its first entry
a condition and its second entry an integer. When the
condition is \code{TRUE}, the competing event variable is simulated
according to the fitted model; when the condition is \code{FALSE},
the competing event variable takes on the value in the
second entry.}

\item{obs_data}{Data on which the model is fit.}
}
\value{
Fitted model for the competing event variable.
}
\description{
This internal function fits a generalized linear model (GLM) for the competing event variable using the observed data.
}
