% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{predict_binomial}
\alias{predict_binomial}
\title{Simulate Binary Values}
\usage{
predict_binomial(x, size, prob)
}
\arguments{
\item{x}{Integer specifying the number of observations to be simulated.}

\item{size}{Integer specifying the number of trials.}

\item{prob}{Numeric vector specifying the probabilities.}
}
\value{
Numeric vector of simulated covariate values under the binomial distribution.
}
\description{
This internal function simulates covariate values from a binomial distribution.
}
