% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{predict_normal}
\alias{predict_normal}
\title{Simulate Normal Values}
\usage{
predict_normal(x, mean, est_sd = NA)
}
\arguments{
\item{x}{Integer specifying the number of observations to be simulated.}

\item{mean}{Numeric scalar specifying the mean of the distribution.}

\item{est_sd}{Numeric scalar specifying the standard deviation of the distribution.}
}
\value{
Numeric vector of simulated covariate values under the normal distribution.
}
\description{
This internal function simulates covariate values from a normal distribution.
}
