% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{predict_trunc_normal}
\alias{predict_trunc_normal}
\title{Simulate Truncated Normal Values}
\usage{
predict_trunc_normal(x, mean, est_sd, a, b)
}
\arguments{
\item{x}{Integer specifying the number of observations to be simulated.}

\item{mean}{Numeric scalar specifying the mean of the distribution.}

\item{est_sd}{Numeric scalar specifying the standard deviation of the distribution.}

\item{a}{Numeric scalar specifying the lower bound of truncation.}

\item{b}{Numeric scalar specifying the upper bound of truncation.}
}
\value{
Numeric vector of simulated covariate values under the truncated normal
                 distribution.
}
\description{
This internal function simulates covariate values from a normal distribution truncated
on one side.
}
