% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{simple_restriction}
\alias{simple_restriction}
\title{Simple Restriction}
\usage{
simple_restriction(newdf, pool, restriction, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{restriction}{List of vectors. Each vector contains as its first entry
the covariate affected by the restriction; its second entry
the condition that must be \code{TRUE} for the covariate to be
modeled; its third entry a function that executes other specific actions
based on the condition (in this case, this function); and its fourth
entry some value used by the function (in this case, the value
the user desires to assign to the covariate when it is not modeled).}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function assists the implementation of a restriction on a covariate in the data
table \code{newdf} by setting lines where the covariate is restricted to a user-specified value.
The data table \code{newdf} is modified in place.
}
