% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate Counterfactual Outcomes Under Intervention}
\usage{
simulate(o, fitcov, fitY, fitD, yrestrictions, compevent_restrictions,
  restrictions, outcome_name, compevent_name, time_name, intvars,
  interventions, histvars, histories, comprisk, ranges, yrange,
  compevent_range, outcome_type, subseed, obs_data, time_points, parallel,
  covnames, covtypes, covparams, covpredict_custom, basecovs, max_visits,
  ...)
}
\arguments{
\item{o}{Integer specifying the index of the current intervention.}

\item{fitcov}{List of model fits for the time-varying covariates.}

\item{fitY}{Model fit for the outcome variable.}

\item{fitD}{Model fit for the competing event variable, if any.}

\item{yrestrictions}{List of vectors. Each vector containins as its first entry
a condition and its second entry an integer. When the
condition is \code{TRUE}, the outcome variable is simulated
according to the fitted model; when the condition is \code{FALSE},
the outcome variable takes on the value in the second entry.}

\item{compevent_restrictions}{List of vectors. Each vector containins as its first entry
a condition and its second entry an integer. When the
condition is \code{TRUE}, the competing event variable is simulated
according to the fitted model; when the condition is \code{FALSE},
the competing event variable takes on the value in the
second entry.}

\item{restrictions}{List of vectors. Each vector contains as its first entry
the covariate affected by the restriction; its second entry
the condition that must be \code{TRUE} for the covariate to be
modeled; its third entry a function that executes other
specific actions based on the condition; and its fourth
entry some value used by the function.}

\item{outcome_name}{Character string specifying the name of the outcome variable in \code{obs_data}.}

\item{compevent_name}{Character string specifying the name of the competing event variable in \code{obs_data}.}

\item{time_name}{Character string specifying the name of the time variable in \code{obs_data}.}

\item{intvars}{Vector of character strings specifying the names of the variables to be intervened
on in each round of the simulation.}

\item{interventions}{List of vectors. Each vector contains a function
implementing a particular intervention, optionally
followed by one or more "intervention values" (i.e.,
integers used to specify the treatment regime).}

\item{histvars}{Vector of character strings specifying the names of the variables for which history functions
are to be applied.}

\item{histories}{Vector of history functions to apply to the variables specified in \code{histvars}.}

\item{comprisk}{Logical scalar indicating the presence of a competing event.}

\item{ranges}{List of vectors. Each vector contains the minimum and
maximum values of one of the covariates in \code{covnames}.}

\item{yrange}{Vector containing the minimum and maximum values of the
outcome variable in the observed dataset.}

\item{compevent_range}{Vector containing the minimum and maximum values of the
competing event variable in the observed dataset.}

\item{outcome_type}{Character string specifying the "type" of the outcome. The possible "types" are: \code{"survival"}, \code{"continuous_eof"}, and \code{"binary_eof"}.}

\item{subseed}{Integer specifying the seed for this simulation.}

\item{obs_data}{Data table containing the observed data.}

\item{time_points}{Number of time points to simulate.}

\item{parallel}{Logical scalar indicating whether to parallelize simulations of
different interventions to multiple cores.}

\item{covnames}{Character string specifying the name of the competing event variable in \code{obs_data}.}

\item{covtypes}{Vector of character strings specifying the "type" of each time-varying covariate included in \code{covnames}. The possible "types" are: \code{"binary"}, \code{"normal"}, \code{"categorical"}, \code{"bounded normal"}, \code{"zero-inflated normal"}, \code{"truncated normal"}, and \code{"absorbing"}.}

\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.). Each vector
must be the same length as \code{covnames} and in the same order.
If a parameter is not required for a certain covariate, it
should be set to \code{NA} at that index.}

\item{covpredict_custom}{Vector containing custom prediction functions for time-varying
covariates that do not fall within the pre-defined covariate types.
It should be in the same order as \code{covnames}. If a custom
prediction function is not required for a particular
covariate, then that index should be set to \code{NA}.}

\item{basecovs}{Vector of character strings specifying the names of baseline covariates in \code{obs_data}.}

\item{max_visits}{A vector of one or more values denoting the maximum number of times
a binary covariate representing a visit process may be missed before
the individual is censored from the data (in the observed data) or
a visit is forced (in the simulated data). Multiple values exist in the
vector when the modeling of more than covariate is attached to a visit
process.}

\item{...}{Other arguments, which are passed to the functions in \code{covpredict_custom}.}
}
\value{
A data table containing simulated data under the specified intervention.
}
\description{
This internal function simulates a new dataset containing covariates, outcome probabilities, competing event
probabilities (if any), outcomes, and competing events (if any) based on an observed
dataset and a user-specified intervention.
}
