% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{threshold}
\alias{threshold}
\title{Threshold Intervention}
\usage{
threshold(newdf, pool, intvar, intvals, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{intvar}{Character string specifying the name of the variable to be intervened on in each round of the simulation.}

\item{intvals}{A list of length 2. The first entry is lower bound of the threshold, and the second entry is the upper bound.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function implements a threshold intervention (i.e., once treatment bypasses a certain
threshold, it remains at that threshold until end of follow-up) for the specified
intervention variable in the data table \code{newdf}. The data table \code{newdf} is modified in place.
}
