% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{carry_forward}
\alias{carry_forward}
\title{Visit Process}
\usage{
carry_forward(newdf, pool, restriction, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{restriction}{List of vectors. Each vector contains as its first entry
the covariate affected by the restriction; its second entry
the condition that must be \code{TRUE} for the covariate to be
modeled; its third entry a function that executes other specific actions
based on the condition (in this case, this function); and its fourth
entry some value used by the function (in this case, the value
the user desires to assign to the covariate when it is not modeled).}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function assists the implemention of a visit process, where a particular covariate is simulated
only when some condition (usually a covariate representing whether a doctor's visit
occurred or not) is \code{TRUE}. If the condition is \code{FALSE}, the covariate value is not
simulated for that time point and the value is instead carried over from the previous
time point. If the condition is \code{FALSE} for more than a certain number of times
consecutively (i.e., if more than a certain number of visits are missed in a row),
the individual is censored.
}
