% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{fit_multinomial}
\alias{fit_multinomial}
\title{Fit Multinomial Model on Covariate}
\usage{
fit_multinomial(covparams, obs_data, j, model_fits)
}
\arguments{
\item{covparams}{List of vectors, where each vector contains information for
one parameter used in the modeling of the time-varying covariates (e.g.,
model statement, family, link function, etc.).}

\item{obs_data}{Data on which the model is fit.}

\item{j}{Integer specifying the index of the covariate.}

\item{model_fits}{Logical scalar indicating whether to return the fitted models. The default is \code{FALSE}.}
}
\value{
Fitted model for the covariate at index \eqn{j}.
}
\description{
This internal function fits a multinomial regression model for a categorical covariate using the observed data.
}
\keyword{internal}
