% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histories.R
\name{lagged}
\alias{lagged}
\alias{cumavg}
\alias{lagavg}
\title{History functions}
\usage{
lagged(pool, histvars, histvals, time_name, t, id_name, baselags)

cumavg(pool, histvars, time_name, t, id_name)

lagavg(pool, histvars, histvals, time_name, t, id_name, baselags)
}
\arguments{
\item{pool}{Data table containing all information prior to time \eqn{t} (\eqn{t} noninclusive).}

\item{histvars}{Vector of character strings specifying the names of the variables for which history functions are to be applied.}

\item{histvals}{For \code{lagged}, this argument is a vector specifying the lags used in the model statements (e.g., if \code{lag1_varname} and \code{lag2_varname} were included in the model statements, this vector would be \code{c(1,2)}). For \code{lagavg}, this argument is a numeric vector specifying the lag averages used in the model statements.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool}.}

\item{t}{Integer specifying the current time index.}

\item{id_name}{Character string specifying the name of the ID variable in \code{pool}.}

\item{baselags}{Logical scalar for specifying the convention used for lagi and lag_cumavgi terms in the model statements when
the current time index, \eqn{t}, is such that \eqn{t < i}. If this argument is set to \code{FALSE}, the value
of all lagi and lag_cumavgi terms in this context are set to 0 (for non-categorical covariates) or the reference
level (for categorical covariates). If this argument is set to \code{TRUE}, the value of lagi and lag_cumavgi terms
are set to their values at time 0. The default is \code{FALSE}.}
}
\description{
These functions create new columns in an input data table for covariate histories. Users must specify which covariates are to be used in the history functions. The data table is modified in place.
}
\details{
\code{lagged} creates new columns for lagged versions of existing
variables in the dataset. The user must specify which variables are to be lagged.

\code{cumavg} creates new columns for the cumulative average up until
time \eqn{t} of existing variables in the dataset.

\code{lagavg} creates new columns for the "lagged cumulative average"
(cumulative average up until time t, then lagged by one time unit) up until time \eqn{t} of existing
variables in the dataset.
}
