% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.gformula_binary_eof}
\alias{print.gformula_binary_eof}
\title{Print method for objects of class "gformula_binary_eof"}
\usage{
\method{print}{gformula_binary_eof}(x, coefficients = FALSE,
  stderrs = FALSE, rmses = FALSE, ...)
}
\arguments{
\item{x}{Object of class "gformula_binary_eof".}

\item{coefficients}{Logical scalar indicating whether to  print the model coefficients. The default is \code{FALSE}.}

\item{stderrs}{Logical scalar indicating whether to print the standard error of the model coefficients. The default is \code{FALSE}.}

\item{rmses}{Logical scalar indicating whether to print the model root mean square errors (RMSEs). The default is \code{FALSE}.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "gformula_binary_eof".
}
\seealso{
\code{\link{gformula_binary_eof}}
}
