% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.gformula_survival}
\alias{print.gformula_survival}
\title{Print method for objects of class "gformula_survival"}
\usage{
\method{print}{gformula_survival}(x, coefficients = FALSE,
  stderrs = FALSE, rmses = FALSE, hazardratio = FALSE, ...)
}
\arguments{
\item{x}{Object of class "gformula_survival".}

\item{coefficients}{Logical scalar indicating whether to print the model coefficients. The default is \code{FALSE}.}

\item{stderrs}{Logical scalar indicating whether to print the standard error of the model coefficients. The default is \code{FALSE}.}

\item{rmses}{Logical scalar indicating whether to print the model root mean square errors (RMSEs). The default is \code{FALSE}.}

\item{hazardratio}{Logical scalar indicating whether to print the hazard ratio between two interventions (if computed). If bootstrapping was used in \code{\link{gformula_survival}}, 95\% confidence intervals will be given. The default is \code{FALSE}.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "gformula_survival".
}
\seealso{
\code{\link{gformula_survival}}
}
