% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{static}
\alias{static}
\title{Static Intervention}
\usage{
static(newdf, pool, intvar, intvals, time_name, t)
}
\arguments{
\item{newdf}{Data table containing the simulated data at time \eqn{t}.}

\item{pool}{Data table containing the simulated data at times before \eqn{t}.}

\item{intvar}{Character string specifying the name of the variable to be intervened on in each round of the simulation.}

\item{intvals}{A list of length 1. The entry is the value of static treatment to be assigned to \code{intvar}.}

\item{time_name}{Character string specifying the name of the time variable in \code{pool} and \code{newdf}.}

\item{t}{Integer specifying the current time index.}
}
\description{
This function implements a static intervention (i.e., either constant treatment or no treatment over
all time points) for the specified intervention variable in the data table \code{newdf}. The data table \code{newdf} is modified in place.
}
